perf: eliminate OOM for 1.5M-node scans
Three memory fixes:
- u128 → u64 for all timestamp fields (modified/created/accessed/scannedAt/elapsedMs):
  saves 24 bytes × node_count (36 MB for 1.5M nodes)
- Drop full ScanResult clone from progress callback: progress ticks now send a
  lightweight {"scanning":true,"nodeCount":N,"elapsedMs":E} line instead of
  cloning ~450 MB and building another ~450 MB JSON string every 3 seconds
- DupeFilter.date_from/date_to narrowed to u64 to match

+++ src/analytics.rs
@@ -19,9 +19,9 @@ pub(crate) struct DupeFilter {
     /// When true, name_pattern must match the whole filename (case-insensitive), not just a substring
     pub(crate) name_exact: bool,
     /// Modified-after threshold in ms since Unix epoch, 0 = no filter
-    pub(crate) date_from: u128,
+    pub(crate) date_from: u64,
     /// Modified-before threshold in ms since Unix epoch, 0 = no filter
-    pub(crate) date_to: u128,
+    pub(crate) date_to: u64,
     /// Path prefix for "original" files (keep side). Empty = no directional filter.
     pub(crate) keep_prefix: String,
     /// Path prefix for "duplicate" files (search side). Empty = no directional filter.
@@ -281,7 +281,7 @@ pub(crate) fn extension_stats(nodes: &[NodeRecord], limit: usize) -> Vec<Extensi
     stats
 }
 
-pub(crate) fn age_stats(nodes: &[NodeRecord], scanned_at_ms: u128) -> Vec<AgeBucket> {
+pub(crate) fn age_stats(nodes: &[NodeRecord], scanned_at_ms: u64) -> Vec<AgeBucket> {
     let mut buckets = vec![
         AgeBucket {
             label: "7 days",
